move modules from lib/ to lib/ui/
