+++ src/lib/db/obtainDb.ts
@@ -1,8 +1,8 @@
 import {to_obtainable} from '@feltcoop/felt/util/obtainable.js';
 import postgres from 'postgres';
 
-import {defaultPostgresOptions} from './postgres.js';
-import {Database} from './Database.js';
+import {defaultPostgresOptions} from '$lib/db/postgres.js';
+import {Database} from '$lib/db/Database.js';
 
 export const obtainDb = to_obtainable(
 	(): Database => new Database({sql: postgres(defaultPostgresOptions)}),
