+++ src/lib/ui/Message_List.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import type {MessageState} from '$lib/messagesStore.js';
+	import type {MessageState} from '$lib/ui/messages_store.js';
 
 	export let messages: MessageState[];
 </script>
