+++ src/lib/ui/socket_store.ts
@@ -1,8 +1,9 @@
 import type {Async_Status} from '@feltcoop/felt';
 import type {Json} from '@feltcoop/felt/util/json.js';
 import {writable} from 'svelte/store';
-import {messages} from '$lib/messagesStore';
-import {posts} from '$lib/postStore';
+
+import {messages} from '$lib/ui/messages_store';
+import {posts} from '$lib/ui/post_store';
 
 // This store wraps a browser `WebSocket` connection with all of the Sveltey goodness.
